Moved user add+delete commands into `utils/users`
